+++ CHANGELOG.md
@@ -4,6 +4,9 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## unreleased
+### Added
+* Added an option to replaces textures with assigned images from current scene for wimdo export.
+
 ### Changed
 * Adjusted wimdo export to report warnings if vertex groups are not part of the original skeleton.